PCA: rename cohort downsampling column param and merge tests
- rename cohort_size_query to cohort_size_column
- keep prepared sample_sets aligned after per-cohort downsampling
- remap selected sample indices when sample sets are reduced
- move downsampling tests into tests/anoph/test_pca.py and delete standalone file
- bump PCA cache key to pca_v9
